feat: moved OnboardingContent to components folder
