+++ react-client/src/routes/my-pets/:petId/onboarding/OnboardingStepOne.tsx
@@ -1,6 +1,6 @@
 import { ASSET_IMAGES } from "~/assets";
+import { OnboardingContent } from "./components/OnboardingContent";
 import { OnboardingPrimaryButton } from "./components/OnboardingPrimaryButton";
-import { OnboardingContent } from "./OnboardingContent";
 import { CommonOnboardingProps } from "./OnboardingDialog";
 import { ONBOARDING_STEPS_TEXTS } from "./onboardingTexts";
