+++ react-client/src/routes/my-pets/:petId/onboarding/components/OnboardingContent.tsx
